+++ docs/CHANGELOG.md
@@ -4,12 +4,15 @@ Changelog
 1.4.0 (Unreleased)
 --------------------------
 - Enh #148: Migration to Bootstrap 5 for HumHub 1.18
+- Enh #152: Use content ID in the "Custom Pages" extension
+- Fix #153: Anonymous badge wrong color
 
-1.3.7 (Unreleased)
---------------------------
+1.3.8 (June 30, 2025)
+---------------------
 - Enh #142: Reduce translation message categories
 - Enh #143: Enable the feature to move a content to another space
 - Enh #145: Updated wording for clarity
+- Enh #151: Extension for module "Custom Pages"
 
 1.3.6 (September 13, 2024)
 --------------------------